backend/pyproject.toml
[tool.poetry]
name = "image-colorizer-backend"
version = "1.0.0"
description = "FastAPI backend for image colorization using DeOldify"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.23"
pillow = "^10.1.0"
torch = "^2.1.0"
torchvision = "^0.16.0"
deoldify = "^0.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"

